requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
executing==1.2.0
Fiona==1.9.4.post1
geopandas==0.13.2
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
install==1.3.5
ipython==8.14.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.6
wcwidth==0.2.6
zipp==3.15.0
